Remove level36.GlActivity from manifest
This line blocks building the project here.

I'm removing it, as there does not even exist
a "GlActivity" class in the level36 package.
If you do have a "GlActivity" class, please use
"svn add" and "svn commit" to add it to version
control, so we don't get build errors here :)