Level 44: Make intro screen broader (to fit the N1 screen ratio)
The Nexus One has a 1.66666666(...) screen ratio with its resolution of
800x480, whereas devices such as the HTC Magic have a 1.5 screen ratio
with the 480x320 resolution.

We work around this problem by making the intro screen broader, which
will fill the screen on the N1 and have the same result on the HTC Magic
(where the overlapping "bars" are simply not visible).